app/backend/pyproject.toml
[tool.poetry]
name = "image-uploader-backend"
version = "0.1.0"
description = ""
authors = ["Maker By Night <2565361+mshafer1@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.1.2"
Flask-WTF = "^1.2.2"
python-decouple = "^3.8"
uWSGI = {version = "^2.0.20", platform = "linux"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
